Clean up photometry subdirectory
